fix execution of rules with command trigger
If a rule is triggered by a received command without specified command
(e.g. "Item <item> received command", so missing optional argument
[<command>]) the rule is not executed.
If a optional command is given "ct.getCommand()" should differ from "null", if
the optional is missing "ct.getCommand()" should be null.
So fix the logic to find a rule (so it is similar to the logic used for
a state update).

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>