add imports used by bundled libs
The bundled xstream library is added to the classpath of the config xml
bundle. So the MANIFEST of xstream is not evaluated by OSGi.
So we have to add the imported packages necessary for the used xstream classes
to the import package section of the config xml bundle.
I added some imports, that are currently used on a ESH instance with the
Yahoo Weather binding running.
I do not know, why this is currently working with Equinox, perhaps
caused by this one:
https://wiki.eclipse.org/index.php/Equinox_Boot_Delegation
But this is something that should be fixed.
Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>