fix thing model handling if factory is not present
If a model for things is added and no thing could be created (e.g. the
factory is missing) the model is not added to the map containing all
thing models.
If a thing handler factory is added the list of things for all models is
processed.
So if a model handles things for a factory, that is not present ATM,
only, the things are not processed later, if the factory has been
registered.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=459648
Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>