change line wrapping policy for enum constants
The previous line wrapping policy for enum constants was set to "Do not
wrap" and the force split was unset, too.
This could result in a non-readable code.
With the new setting "Wrap all elements, every element on a new line"
and the activated "force split" every enum value gets its own line.
So, if you write stuff e.g.
public enum SomeEnum {
    VALUE1(paramA, paramBcde),
    VALUE2(paramCde, paramDefghi),
    VALUE3(paramEfghiklmn, paramF),
    ...
}
it keeps readable and will not changed in
public enum SomeEnum {
    VALUE1(paramA, paramBcde), VALUE2(paramCde, paramDefghi),
    VALUE3(paramEfghiklmn, paramF), ...
}
anymore.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>