console.fw.karaf: add simple implementation for Karaf
We add the dependency (Karaf console core) to the pom file as long as it
is not present in the target platform of ESH.
So we could build the project using Maven + Tycho (as long as
pom dependencies are considered it set to true). The build in the
Eclipse IDE is still not working.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>