improve testing
* remove Require-Bundle in favor of Import-Package
* bump Tycho
* use new dependency resolution for Tycho
* for Tycho bump there are two things that are improved / solved for
  testing:
  * https://wiki.eclipse.org/Tycho/Release_Notes/0.21
    (INCOMPATIBLE CHANGE) In case you have used test-runtime only
    dependencies in the tycho-surefire-plugin configuration, you may need
    to replace them with
    target-platform-configuration/dependency-resolution/extraRequirements
    instead. See bug 436617 comment 11 for details.
  * https://wiki.eclipse.org/Tycho/Release_Notes/0.22
    Tycho Surefire was updated to use Maven Surefire version 2.17. This
    enabled us to consume the fix for long-standing bug 369266
* use JUnit 4 and drop JUnit <4 usage

If you are using "mvn clean install" all went fine.
To get the build also working (all the time) using "mvn install"
(without clean), we have to unset the environments.
If the environment is still set for tests, the test will fail, because
some requirements for the SWTs are incorrect (sure, we cannot use linux,
MAC OS X and windows at the same time).
Why this is necessary for "install" without clean is out of scope of
this commit.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=476024
Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>