avoid ConcurrentModificationException
The functions called for the resources will result on a change of the
resource set (at least on my special system setup).
Using the logic of this PR without using a copy of the resource set
(use getResources() for the for each loop) will result in a CME, too.
I cannot reproduce my CME using the copy.
As long as there is no Xtext guru, I cannot identify if this copy fix
the problem for all possible timing cases.

This PR also remove the usage of some Guava code, that is not really
necessary (we do not need to create two iterators).

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=479215
Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>