channel group id can contain illegal item chars
If we use a channel group type multiple times in a thing type, we assign
different channel group ids to them.
Let's consider a thing that is using the channel group type "magic"
twice.

...
<channel-groups>
    <channel-group id="magic-1" typeId="magic" />
    <channel-group id="magic-2" typeId="magic" />
</channel-groups>
...

A hyphen is a valid character for an ID, but it is not a valid character
for an item name.
If we are using the channel-group ID to generate an item name, we have
to call toItemName also for the ID.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>