paper ui: thing cfg page miss channels of groups
The getChannels function iterates over all channels of a thing.
For every channel the channel it uses the getChannelTypeById function
to get a channel object supplied by the thing type.
This channel object is then further processed.
The getChannelTypeById looks for the given channel by the supplied
channel id in the channel array of the thing type only.
So all channels that are part of a channel group of the thing type are
missing.
We have to also look for the channel in the channel groups of the thing
type.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>