#659 - Philips Hue Binding: unsupported light
Changed Thing usage in that way that it is not anymore necessary to
include specific bulb types as single Thing-Types. From now on there are
4 generic thing types that are supported (Dimmable light, Color light,
Extended Color light, Color temperature light) with there respective
channels. The Thing type to use for a discovered bulb is now determined
by it's type (see
http://www.developers.meethue.com/documentation/supported-lights)

https://github.com/eclipse/smarthome/issues/659
Signed-off-by: markusmazurczak <coding@markus-mazurczak.de>