fix timing dependend test execution
There are currently two tests in the ItemUpdateOSGiTest.
* ItemUpdater sets item state
* ItemUpdater sends state changed event

The "sets item state" test triggers a changed event
* UnDefType.NULL => OnOffType.ON

The "sends state changed event" test triggers a changed event
* OnOffType.ON => OnOffType.OFF

The first test does not check / reset the event.
The second test checks for the ON => OFF event.
If the first event is stored to "last event" but the second on is not
yet received, the check for "ON => OFF" will fail.

The receiving of the event is done asynchron, so we have to use correct
sync mechanism, too.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>