Removing bower
NOTES:

tinycolor2 is the equilivent of tinycolor bower file, https://github.com/bgrins/TinyColor#including-in-node
Latest version 4.3 of roboto-fontface was never published to NPM, package no longer seems to be maintained. https://github.com/choffmeister/roboto-fontface-bower, https://github.com/choffmeister/roboto-fontface-bower/issues/25
angular-material 10.0 has a peer dependency on angular version 1.3.15 or higher. I moved to angular 1.3.15, was never an issue with bower because we were not hard referencing versions.

angular-animate was never included in the bower file but is a peer dependency of angular-material. angular-animate is being used in the app
angular-aria was never included in the bower file but is a peer dependency of angular-material. angular-aria is being used in the app

My guess is with the two peer dependencies is that bower auto installed them.

Signed-off-by: Grady Duncan <gradyduncan@gmail.com>