fix ThingSetupManager removal handling
The ThingSetupManager used to listen to ThingRemovedEvents, which are sent out not only if a Thing gets removed explicitly, but also if a ThingProvider gets removed (the ThingRegistry notifies the listeners by sending out the event). As a result, the TSM deletes will also delete Items etc. if during shutdown DS deactivates the ManagedThingProvider before the ThingRegistry, leading to data loss (i.e. room assignment and labels cannot be restored).

Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>