automation timer test stabilization
The current test sets an item, check if the state of the item is the set
one and then checks if the state has been changed to another one. The
state change is triggered by a timer triggered rule.
The timer activates the rule every second (using wildcard for the cron
expression statements).
If the rule is triggered between the set state and the assertion that
expects the state is the set one, the test fails.
This commit changes the logic that it disables the rule to check if the
value has been set and then enables it again afterwards before checking
if the rule has been applied.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>