HueBridgeHandler: Fix isEqual(state,state) Method
Before, every Hue light state which was checked for equality resulted in true (=equal) if the light did not support color or alert.
This is not the full truth: Thanks to the `&&` expression, the isEqual() order of equal checking was:

 1. alert
 2. isOn
 3. effect
 4. brightness
 5. color mode
 6. ...

So it only fails if nothing before color mode was not equals.

Signed-off-by: Sebastian Janzen <sebastian@janzen.it>