add a base64 encode / decode wrapper for Java 7 and 8
We could use a bundle that wraps some methods that real implementation
differs if we use Java 7 or Java 8 as the platform JRE.
We do not need to load wrappers for multiple JREs if we need only one at
runtime. Also we do not need to use dynamic imports or some other stuff
that could trigger CNF exceptions on runtime.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>