Do not cut of script exception
This is related to:
https://github.com/eclipse/smarthome/issues/1393#issuecomment-211540325
Which message should be shown, the message of the exception themselves
or the message of the exception that is the cause of the excpetion.
I do not see any reason to handle all the different cases here and
generate a suitable message itself.
We should not limit the information of the exception for logging.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>