fix jersey servlet support
If the jersey modules are initiated on the same level as the ones that
registered the ServletContainer class, this could lead to errors using
asyncrhonous processing:
javax.servlet.ServletException: java.lang.UnsupportedOperationException:
Asynchronous processing not supported on Servlet 2.x container.

See e.g.
https://gist.github.com/maggu2810/ef57c2656fa9f76052042da5f70d9d2f

This could be worked around to start the Jersey components in front of
the usage.
The important module is jersey-core-server, that adds support for
servlet 3.x.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>