fix minimum / maximum text (#1496)
If the minimum / maximim is set to some value that evaluates to false
(e.g. 0, 0.0, ...) then the minimum / maximum text is not shown.
The slider element is shown only if the minimum and maximum text
evaluates to true.
So with the current test no slider is shown if the min / max is e.g. 0,
0.0, ...

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>