Sonos binding: fix issues #1900 and #1902
Player UI actions are now working when applied to a slave member of a group.
Implement exception for missing zone player handler
Fix NPE when no thing is defined for the group coordinator
Change of state is fixed for GroupCoordinatorIsLocal, MuteMaster, LineInConnected and AlarmRunning.
saveAllPlayerState andrestoreAllPlayerState are now considering really all Sonos players.
Deprecated calls to updateState have been updated.
Fix: method onValueReceived: add a missing break in the java switch
Fix: method updateSleepTimerDuration: onValueRecived was called with service "DeviceProperties" rather than "AVTransport"
Enhancement: add a warning message if the handler for a zone player is not found (when searching the handler of the group coordinator)
Enhancement: quick refresh of the zone name without waiting for the polling job
Enhancement: quick refresh of the sleep timer when updated from the Sonos application (without waiting for the polling job)
Enhancement: quick refresh of the LED state when updated from Eclipse Smarthome / openHAB

Signed-off-by: Laurent Garnier <lg.hc@free.fr>