When creating Things, keep their Channels in order
The `Thing` interface declares its channels property as `List<Channel>`,
implying that the order of Channels on a Thing is stable.
This property is kept by the rest of the code, all of it,
except for `ThingResource#create`.

When a `Thing` is created via the REST interface, the `Channel`s
specified by the request are collected in a `HashSet`,  shuffling them.

This change collects the `Channel`s in an `ArrayList` instead,
guaranteeing that the order specified in the request is kept
and that `Channel`s aren't randomly shuffled.

Signed-off-by: Martin Kühl <martin.kuehl@gmail.com>