Cleanup discovered children if a Bridge Thing was removed
When DiscoveryResults had been found via a Bridge, and this Bridge
was removed, so far these DiscoveryResults remained in the inbox.
However, as we assume that a Thing having a Bridge actually requires
it in order to communicate successfully, it does not make sense
anymore to keep it in the inbox without the Bridge. There won't be
any useful result when approving it, therefore we easily can clean
them up.

fixes #1502
Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>