Drop Jetty v8 in launch configuration
This PR removes the usage of Jetty v8 in the Eclipse SmartHome Runtime
launch configuration.
It also removes unneeded p2 bundles from the launch config etc.
At least it renames the launch configuration from "SmartHome Runtime" to
"SmartHome-Runtime" because the launch configuration's name is used as
working directory for the OSGi container and whitespaces should be
avoided (at least for the jetty home bundle path).

Fixes: https://github.com/eclipse/smarthome/issues/2032
Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>