add / remove provider logic requires active component
After some modifications of the ItemChannelLinkRegistry yesterday (which
have been never going upstream) I realized that the AbstractRegistry
implementation relies on an active component if provider are added (and
removed).
Using the multiple (0..n) cardinality and policy static the providers
could be added / removed if the component is active but also if the
component is not active.
Regardless if the error is currently thrown this is an error in the
design and should be fixed.

This PR introduce a service tracker to track the providers as long as
the component is active.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>