Add test to ensure that group functions persist
With 1cf093e we can create group items with group functions via the REST API, but the group functions are not currently persisted: ManagedItemProvider.PersistedItem is missing affordances to store group functions and ManagedItemProvider is missing code making use of them.

This change adds a test to ensure that group functions are saved and restored correctly. The test is currently failing but might benefit the future implementer.

Signed-off-by: Martin Kühl <martin.kuehl@gmail.com>