Added setPercentageProgress(int progress, boolean silent) to ProgressCallback Added progress to FirmwareUpdateProgressInfo
setPercentageProgress allows users of the ProgressCallback to set the percentage progress
eigther silent (no event posted) or not silent (event is posted if progress changed).
If silent or not, the updated progress will be part of the next FirmwareUpdateProgressInfo Event (for example if the ProressStep changes).
This allows update processes driven by ProgressStep changes or percentage update progress changes.

Signed-off-by: Christoph Knauf  <christoph.knauf@itemis.de>