Changes in the FSInternetRadio binding.
Some changes are made in org.eclipse.smarthome.bindig.fsinternetradio.
They are mainly related to cleaning up some unneccessary checks and
code lines that can never be reached.
The HTTP client now is stopped with the deactivation of the bundle.
The hardcoded "fsapi" String is now extracted into an instance variable.

Signed-off-by: Mihaela Memova <mihaela.memova@musala.com>