Changed AudioHTTPServer to return relative urls instead of absolute
This moves the responsibility of constructing the absolute url to call to the consumer - which is imho the much better approach, as e.g. a consumer like a web browser might request it remotely (from a different network) and thus only it knows which server+port to contact (and with this to avoid any cross origin problems).

- adapted Sonos binding to make relative url absolute
- added configuration parameter for callback url to Sonos binding (if not set, heuristic approach through scanning available network interfaces is taken)
- tested it also successfully with webaudio sink in Paper UI

Signed-off-by: Kai Kreuzer <kai@openhab.org>