state that null is an allowed return value
The current JavaDoc states, that a map of values should be returned.
The current implementations of that interfaces use null as a return
value, too.
Also the rule engine checks the return value and checks for null.

I personally prefer to state that the function must not return a null
value and should return an empty map instead of adding a null check in
the handling of the return value.
But if we don't want to break third party action implementations we
should at least state, that null is an allowed return value.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>