ThingHandler shutdown symmetric to initialization
Initialization of ThingHandlers happens in the following order:
* Thing -> INITIALIZING
* BaseThingHandler gets the callback injected
* ThingHandler.initialize()
* Thing -> ONLINE/OFFLINE/UNKNOWN

The shudown however so far first removed the callback, and then called
ThingHandler.destroy(). As a result, the ThingHandler did not get the
chance to clean up first and could run into situations where it
tried to e.g. send events but the callback was already missing.

Hence the shudown process with this change becomes the exact reverse:
* Thing -> UNITIALIZED (so that it's not used anymore)
* ThingHandler.destroy()
* BaseThingHandler gets the callback "un"injected
* Thing -> UNINITIALIZED (HANDLER_MISSING)

fixes #2757
fixes #1338
Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>