Fix issues in thing update/change rules.
The rule is changed to take thingUid as a string instead of an ID. And we
don't use the type Thing internally to pass the information in the rule. This
is done so because when the thing is removed, it can occur before the rule is
checked and executed. When it occurs, we cannot get the Thing from thingUid.
Fix an issue in storing thing related rules and retrieve them.

Bug: https://github.com/eclipse/smarthome/issues/1654
Signed-off-by: Maoliang Huang <hellomao@outlook.com>