DBus bundle: allow usage of exported libraries in Eclipse IDE (#3074)
The content of 'lib/unix-0.5.jar' and 'lib/libdbus-java-2.7.jar' is
already added to the bundle classpath and packages are already exported
by OSGi.
The Eclipse IDE classpath settings does not mark that stuff of that
files is exported. So access to e.g. org.freedesktop.dbus is not
available in the Eclipse IDE.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>