OSGi test: return implementation class
If we get a service class by filter on a specific implementation class
an instance of the implementation class (so the more specified) should
be returned.
This is what is currently done and used.
It seems Groovy ignores the types and does not warn / error if you
assign the returned serice class to an instance of the implementation
class.
If you use the Java compiler it errors as expected.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>