fix race condition which could leave a thing in INITIALIZING (#3088)
as we have seen in #2945 and #3015, there can be situations when a thing
is left in INITIALIZING state although the binding updated its state
correctly.

The root cause was a race condition between ThingManager and GenericThingProvider.
While the ThingManager was trying to initialize the Thing, the GenericThingProvider
tried to update the the existing thing (which already had a ThingHandler). As a
result, the existing ThingHandler updated the outdated thing, not knowing that
there actually was an update.

The key for fixing this was changing the point in time when a handler is created.
This will now also be postponed until the binding is fully loaded. Additionally,
bindings have to deal with thing updates while the thing still is in INITIALIZING
state. On the other hand, it is now technically prevented that ThingHandler.initialize()
can get called again when a thing already is ONLINE/OFFLINE/UNKNOWN.

fixes #2945
fixes #3015
Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>