prevent bindings from removing things
...and making the documentation more explicit about this.

So far, bindings simply could set things to REMOVED at any time so they got
dropped from the database. Now it is ensured they can only do so after the
framework set it to REMOVING, i.e. they only are allowed to indicate that
the handleRemove() is finished, but cannot trigger the process anymore.

fixes #3199
Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>