fix FSInternetRadioHandlerOSGiTest
as in the "verify OFFLINE Thing-status and NULL Item-state when the HTTP
response cannot be parsed correctly" the property to return invalid XML
was set too late. When timing was right, the handler hat the chance to
set the item state once, therefore the assumption that it was not set
failed.

Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>