use classic scheduler for one-time jobs
The (cron) expression based scheduler is still used for the "daily" jobs
which will be executed every midnight. Those "daily" jobs however create
jobs for all kinds of events during the current day, which will have to
be executed only once. For those, it does not really make sense to convert
a timestamp to a cron expression, which the scheduler again with lots of
effort boils down again to a timestamp. Therefore they are now simply
executed with a calculated delay directly using the regular scheduler.

Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>