fix number deserialization on Json Storage
As it turned out, the current Map deserializer was not very effective
in converting numbers to BigDecimals as it took itself out of the game
when deserializing the outer map which JsonStorage adds itself. Hence
a property map within an entity lost its normalization.

This could have been compensated again during read by simply cutting off
the trailing ".0"s, but this still led to the awkward situation that the
numbers in the serialized format were altered and looked different of
what would have been expected.

Therefore this change
* Distinguishes the "outer map" (i.e. the internal data structure of the
  JsonStorage and the inner entity structures
* Uses two different Gson instances for handling those two structures
  individually
* Lets JsonStorage internally keep JsonElement structures _always_
* Registers a custom deserializer for Configuration which handles the
  custom number-to-BigDecimal conversion

fixes #3774
Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>