Add new binding for LIRC. (#2537)
* Add new binding for LIRC.

The binding does not send the number of repeats and does
not support UNIX sockets, so the LIRC server must be
exposed as over TCP.

* [LIRC] Change how repeats are handled

The way LIRC handles transmitting repeats isn't how most users
expect it to behave. Sending the command "KEY_VOLUMEUP 5" would
actually transmit the volume up code six times. The commit
changes that behavior to be more user-friendly. This also
includes examples on using the binding.

Signed-off-by: Andrew Nagle <kabili@zyrenth.com>