MqttGeneric binding
Bind Mqtt topics to things. Broker connections are listed as brigdes. A generic thing
can be added. The thing can host one or multiple channels, each linked to a mqtt state / command
topic and an own optional transformation. A channel is one of the following types:

* TextChannel
* NumberChannel
* PercentageChannel
* OnOffChannel

A channel can be configured with a transformation pattern
for an incoming message from a Mqtt state topic. This allows to extract a value from
JSON/XML/etc, as it seems to get popular on some Mqtt topics to encode multiple values into json.

The handlers and all value classes are backed by tests.

Signed-off-by: David Gräff <david.graeff@web.de>