Mqtt Broker Connections binding
This is the main mqtt binding.

At the moment it is possible to configure brokers in a textual way. Without intensively watching the log file, there is no way to tell if the provided configuration is correct and if the connection to the broker have been established, neither via the GUI (PaperUI) nor machine readable (REST interface).

This binding makes the configuration of brokers more intuitive and let the user add broker connections as things.
The connections are synced to the io.transport.mqtt MqttService and are therefore available for all other bindings to use.

Features include:
* 2 different Bridges: SystemBrokerConnection, BrokerConnection
* Rich status feedback for textual configured brokers. No longer skimming through log files if a connection doesn't work.
* Auto discovery of brokers (Enumeration of MqttService shared broker connections, best effort subnet and localhost scan on standard Mqtt ports)
* Configurable reconnect time (this is not possible via textual configuration yet)
* Certificate and public key pinning for ssl connections to brokers
* Extensive tests for all handlers, the discovery services and the SSLContextProvider.

Signed-off-by: David Graeff <david.graeff@web.de>