MqttGeneric binding
Bind Mqtt topics to things. Broker connections are listed as brigdes. A generic thing
can be added. The thing can host one or multiple channels, each linked to a mqtt state / command
topic and an own optional transformation. A channel is one of the following types:

* TextChannel
* NumberChannel
* PercentageChannel
* OnOffChannel

A channel can be configured with a transformation pattern
for an incoming message from a Mqtt state topic. This allows to extract a value from
JSON/XML/etc.

Auto discovery is also supported. The following two conventions are implemented:
* Homie 3.x MQTT convention (except node instances).
* HomeAssistant MQTT components (Fan, Cover, Climate, Camera are stubs)

The handlers and all value classes are covered by tests.

Signed-off-by: David Graeff <david.graeff@web.de>
Signed-off-by: David Gräff <david.graeff@web.de>
Also-By: Simon Kaufmann <simon.kfm@googlemail.com>
Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>