Respond with status code 404 for missing icons
If an Icon does not exist, the request requesting this icon should respond
with a 404 status code and let the client decide, how it wants to handle
such cases.

In this commit, the IconServlet is changed to do not automatically deliver
the none.png image when an icon does not exist. The basicui and classicui
are updated (as they're clients for this servlet) to automatically load
none.png whenever the img tag of an icon fails to load. The behaviour
visible for a user is the same.

Fixes #3842

Signed-off-by: Florian Schmidt <florian.schmidt.welzow@t-online.de>