Remove check when updating in the AbstractRegistry that was introduced when merging ESH changes in Sprint 148.
The removed checks prevent the generation of ItemUpdatedEvents when
items retrieved from the ItemRegistry are modified, like in the
following snippet:

Item item = itemRegistry.get(itemName);
item.setLabel("blabla");
itemRegistry.update(item);

The underlying reason is that, currently, the item registry is not
returning copies of the items, but the reference to the item managed by
the registry. This shall be changed in ESH, the corresponding bug ticket
is here: https://github.com/eclipse/smarthome/issues/4009

Once the ESH bug is resolved, we should re-add the checks that are
removed by this commit.