fix potential CME
The items member are used internally as expected and additionally by the
internal item registry change listener.
The items are iterated and the map could be modified at the same time.

The getAll function currently returns a view of the map values. So, the
caller could run into trouble using the returned set if the internal map
is changed.

Fixes: https://github.com/eclipse/smarthome/issues/3956
Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>