storage: add containsKey method
The storage interface returns null if you call get for a key that is not
present in the storage.
If you want to check if a key is present (and differ between non-present
or present but an assigned null value) you currently need to call
`storage.getKeys().contains(key)`.
Normally there exist better implementation that does not require to
create a collection of all keys first and then check that collection.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>