Core/config: Add NestedConfigConverter
The added class converts between a flat map of <key,value> tuples to a nested map and back.
`key`s in the flat map express a hierarchy by the hash sign and dots. The schema is:

key.instanceName#key.instanceName#...#key=value

The `instanceName` can be anything but need to be unique among all instances at the same
hierarchy level for a given key.
The `key` part identifies the field name for the nested structure.

The pairs of `key`.`instanceName`, separated by a dot, that form a hierarchy are separated by dashes.

<h2>Example</h2>
A flat map looks like this:

autostart=true
broker.broker1#host=123.123.123.123
broker.broker1#secure=true
broker.broker2#host=124.124.124.124
broker.broker2#secure=true

and will be converted to:

autostart=true
broker
   broker1
     host=123.123.123.123
     secure=true
   broker2
     host=124.124.124.124
     secure=true

This class is useful in conjunction with {@link Configuration} and its
(un-)marshaling functionality.

Signed-off-by: David Gräff <david.graeff@web.de>