Refactored OSGi/Guice integration of DSL based models
A new ServiceModule is introduced which takes care to inject ESH
services via Guice in order to prevent occasional
"ScriptServiceUtil is not initialized" errors.

Also, the EMF registries are cleaned up if model parsers are
stopped and their corresponding resources are removed from
the ModelRepository. This is to prevent class cast execptions
when old instances hit a newer version of the rule engine.

fixes #3157
Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>