Leverage link configuration to select the profile
With this, the ItemChannelLink configuration can be used to select
which profile should be used:

Switch someItem {
      channel="a:b:c:channel1" [
          profile="master"
      ]
      channel="a:b:c:channel2" [
          profile="system:slave"
      ]
      channel="a:b:c:channel3" [
          profile="mybinding:foo"
      ]
}

The profile identifiers are namespaced, so bindings can bring their own.
There is a "system" namespace which also is the default if no namespace is
given (like in the "master" example above).

There is a config description available for the links under

    link:<itemName> -> <channelUID>

which also includes all applicable profiles in its options.

fixes #2226
Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>