[BasicUI] Sliders handle NumberItems gracefully
The forceful conversion of any state to a PercentType has shown
to be desastrious when having Sliders represent Number items.
Although they are primarily intended to be used with Dimmer items
only, this is the only place which prevents them being used for
Number items, causing an Exception which was pretty confusing for
users.

fixes #4397
Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>