WIP: External null annotations via zip
This PR is similar to #4217 but instead of creating a project that
contains the annotations in a directory it sets up the IDE to read the
annotations from a zip file.

The reason behind this is that the IDE should be for the every day
developer who will not create external annotations. Nevertheless he
should benefit from them which is why they are provided in a zip file.

This zip file should be hosted externally, maybe together with
lastnpe.org. That is also the place to add new annotations, i.e.
download the project from there, change the IDE to use this project as a
directory, add your annotations and afterwards the project releases a
new zip file.

In this PR I created a setup task for the IDE to download the
annotations file and place it in a "eea" directory in the targetplatform
folder. For maven there is a new dependency to the wagon-maven-plugin
which enables downloading a single file to be placed where the IDE
stores the file too. The compile task uses this zip file as
"annotationpath".

Also I reactivated the "Unchecked conversion from non-annotated type to
@nonnull type" check by setting it to warning level, otherwise (if set
to ignore) external annotations are not needed.

As an example I updated the ".classpath" file of the sonos binding to
use the downloaded zip file.

The parts which are still valid from PR #4217 are:
* all created external annotations
* all updated .classpath files (IF adjusted to use the zip file instead
of a directory)

The URL where the file will be hosted is still left open, which is why
this PR is WIP.

Signed-off-by: Stefan Triller <stefan.triller@telekom.de>