WIP: Provide information about multiple instance services for UIs
The newly introduced MultipleInstanceServiceInfo interface serves as a
marker service for the service pid returned by its getServicePID()
method.

Every service that should be configured multiple times with different
configurations should have the configurationPolicy
ConfigurationPolicy.REQUIRE and thus it will not be started without a
proper configuration.

In addition one service implementing MultipleInstanceServiceInfo is
neccessary that is started immediately which announces the availability
of the forementioned-service by returning the configurationPid of it in
its getServicePID() method.

Signed-off-by: Stefan Triller <stefan.triller@telekom.de>