On adding Thing from Inbox check for missing settings
After adding a Thing from Inbox an additional check is performed:
- is the Thing OFFLINE and has CONFIGURATION_ERROR _or_
- is the Thing _not_ ONLINE and has some empty configuration entries
If so, user will be redirected to the Thing Edit page automatically.

This implementation present a seamless experience for a user, for
example allowing to discover a bridge, be prompted to enter its
credentials and to discover connected Things in a same workflow.

Small drawback: if a Thing takes long to become ONLINE _and_ has some
empty non-mandatory configuration entries, user still will be
redirected.

To reduce this risk, this method could be applied to bridges only,
feedback is welcome.

Signed-off-by: Pavion <priakhine@yahoo.de>