extend HSBType
The HSBType is useful for dimming with the same color, because it is easy to change the brightness channel. However, the HSBType currently only has get, not set methods for hue, saturation and brightness. So changing the brightness requires the creation of a new object. This PR adds the missing set methods.

Signed-off-by: Jan N. Klug <jan.n.klug@rub.de>