Changes worked in!
I have no idea how or why I came up with that "NotificationSoundVolume" :blush:. 

Regarding the Validation error, I posted initially without the "(github.JueBag)" , because it is/was working that way on the OH side. I added it this part because I saw the error. This is my first one in here! 

Signed-off-by: Jürgen Baginski opus42@gmx.de (github:JueBag)