align WatchService events on all platforms/file systems
Apparently some platforms signal for every file creation/deletion
a modification of the containing directory while others dont.

As this appears to be not only specific to some OS but also
depends on the file system used, such directory modifications
are suppressed now completely.

So instead of adding yet another case to the test I rather
streamlined the behavior of the productive code to behave
consistently (and thereby ported the test-case to Java).

fixes #3952
Signed-off-by: Simon Kaufmann <simon.kfm@googlemail.com>